requirements.txt
# Core Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# CORS Support
flask-cors==6.0.1

# Database
psycopg2-binary==2.9.7

# Environment Variables
python-dotenv==1.0.0

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Audio Processing (optional)
librosa==0.10.1
soundfile==0.12.1

# Machine Learning (optional)
scikit-learn==1.3.0

# HTTP Requests
requests==2.31.0

# Development Tools (optional)
flake8==6.0.0
black==23.7.0
pytest==7.4.0